+++ modules/castbar.lua
@@ -337,6 +337,7 @@ function BBP.CustomizeCastbar(frame, unitToken, event)
                     local start, duration = BBP.TWWGetSpellCooldown(interruptSpellIDx)
                     local cooldownRemaining = start + duration - GetTime()
                     local castRemaining = (endTime/1000) - GetTime()
+                    local totalCastTime = (endTime / 1000) - (castStart / 1000)
 
                     if not notInterruptible then
                         if cooldownRemaining > 0 and cooldownRemaining > castRemaining then
@@ -349,13 +350,60 @@ function BBP.CustomizeCastbar(frame, unitToken, event)
                                 castBarTexture:SetDesaturated(true)
                             end
                             castBar:SetStatusBarColor(unpack(castBarDelayedInterruptColor))
+
+                            if cooldownRemaining < castRemaining then
+
+                                if not castBar.spark then
+                                    castBar.spark = castBar:CreateTexture(nil, "OVERLAY")
+                                    castBar.spark:SetTexture("Interface\\CastingBar\\UI-CastingBar-Spark")
+                                    castBar.spark:SetSize(16, 52)
+                                    castBar.spark:SetBlendMode("ADD")
+                                    castBar.spark:SetVertexColor(0,1,0)
+                                end
+
+                                -- Calculate the position of the spark (interrupt percentage)
+                                local interruptPercent = (totalCastTime - castRemaining + cooldownRemaining) / totalCastTime
+
+                                -- Adjust the spark position based on the percentage of the cast bar
+                                local sparkPosition = interruptPercent * castBar:GetWidth()
+                                castBar.spark:SetPoint("CENTER", castBar, "LEFT", sparkPosition, -2)
+                                castBar.spark:Show()
+
+                                -- Schedule the color update for when the interrupt will be ready
+                                C_Timer.After(cooldownRemaining, function()
+                                    if castBar then
+                                        if not castBarRecolor and not useCustomCastbarTexture then
+                                            if castBarTexture then
+                                                castBarTexture:SetDesaturated(false)
+                                            end
+                                            if not classicFrames then
+                                                castBar:SetStatusBarColor(1, 1, 1)
+                                            end
+                                        else
+                                            if casting then
+                                                castBar:SetStatusBarColor(unpack(castBarCastColor))
+                                            elseif channeling then
+                                                castBar:SetStatusBarColor(unpack(castBarChanneledColor))
+                                            end
+                                        end
+                                        -- Hide the spark once the interrupt is ready
+                                        if castBar.spark then
+                                            castBar.spark:Hide()
+                                        end
+                                    end
+                                end)
+                            else
+                                if castBar.spark then
+                                    castBar.spark:Hide()
+                                end
+                            end
                         else
                             if not castBarRecolor and not useCustomCastbarTexture then
                                 if castBarTexture then
                                     castBarTexture:SetDesaturated(false)
                                 end
                                 if not classicFrames then
-                                    castBar:SetStatusBarColor(1,1,1)
+                                    castBar:SetStatusBarColor(1, 1, 1)
                                 end
                             else
                                 if casting then
@@ -363,6 +411,9 @@ function BBP.CustomizeCastbar(frame, unitToken, event)
                                 elseif channeling then
                                     castBar:SetStatusBarColor(unpack(castBarChanneledColor))
                                 end
+                            end
+                            if castBar.spark then
+                                castBar.spark:Hide()
                             end
                         end
                     end